fixes -
ST-39: In Cashier the line: this.sector = new SimpleObjectProperty<Sector>((Sector) inputStream.readObject()); writing Sector inside <> is redundant as Java can already infer the generic type
ST-40: In Cashier Importing “javafx.collections.FXCollections” was unnecessary, also importing “java.util.Date” and “java.util.List” was unnecessary as we are using the “time” library
ST-49: In the declaration of the method deleteAll, in CategoryFileHandler,  ArrayList was replaced with List so we do not have confined behaviour. List is better as it can be a place holder for ArrayList, LinkedList etc. The line “import java.util.ArrayList;” was removed as it is no longer needed and “import java.util.List;” was added.
ST-50: In CategoryFileHandler, unused exception variables in catch blocks were removed and replaced with unnamed patterns, as the exceptions were intentionally not used. Comments were added to document that these blocks represent normal control flow (e.g., end-of-file conditions), improving code clarity and maintainability.
ST-82: Several unused imports inside InventoryController were removed
ST-83 and ST-94 and ST-104: 6 data fields of type string were added in InventoryController, all final. They each hold significantly repeated text. Each usage of these texts was replaced with the corresponding variable.
ST-84: Useless curly braces were removed in InventoryController.
ST-85: A few unnecessary “.toString()” were removed in InventoryController as the method already returned a string.
ST-86 and ST93 and ST103: In InventoryController all static methods are now accessed using the class name instead of an instance reference to improve clarity and correctly reflect class-level behavior.
ST-87: In InventoryController inside the sendAlertforLowStock() method the flow of the if else statement led to a redundant return statement. Now in the if block we check if the lowStockItems is not empty then we continue with the block, else the method terminates.
ST-89: In InventoryController there was a block of code that was commented out, it is now removed.
ST-95 and ST-105: In InventoryController In the constructor, the parameter name currentUser was changed to user to avoid hiding the class field currentUser. Although the assignment using this.currentUser was correct, using the same name for both the constructor parameter and the instance variable caused variable shadowing. Renaming the parameter improves code clarity, prevents ambiguity, and aligns the implementation with SonarQube and clean code best practices.
ST-120: In LoginController, there was a chunk of commented out unnecessary code that was removed. Also there was an unnecessary usage of curly brackets, they were removed. Unnecessary imports were also removed.